// packages/channel-yuanbao/src/access/ws/types.ts
export type WsSyncCommand = {
  name: string;
  description: string;
};

/** SyncInformation command data (sync_type=1) */
export type WsSyncCommandsData = {
  botCommands: WsSyncCommand[];
  pluginCommands: WsSyncCommand[];
};

/** SyncInformation request data */
export type WsSyncInformationData = {
  syncType: number;
  botVersion: string;
  pluginVersion: string;
  commandData?: WsSyncCommandsData;
};